Generator tool: Now reads JavaDoc and parameter names.
The generated syntactic sugar (e.g. Matchers, CoreMathers) was previously based purely on reflection which meant it couldn't read the JavaDoc and parameter names.

This meant the resulting sugar class didn't contain enough information for users, forcing them to have to dig deeper.

This change uses QDox to read the original Java source code of the matchers meaning the final generated class has everything necessary for the user to understand it. 

If the source code cannot be read, the generator still proceeds, just without this extra info.