Escape backslashes properly in BaseDescription
Before this commit, the following assertion:

```
        assertThat("\n", equalTo("\\n"));
```

gives the output:

```
java.lang.AssertionError:
Expected: "\n"
     but: was "\n"
```

This is confusing because it seems like two identical strings have not
compared equal!

The problem is that in BaseDescription.toJavaSyntax(char) escapes
newline characters but doesn't escape backslash characters, so the two
strings being compared are transformed to the same output for display,
despite having different actual content.