Merge core and library, revive hamcrest-integration (#229)
This is a big change to the project structure, but there are no changes to any Java code.

As this is a breaking change due to the repackaging, the version has changed from 1.4-SNAPSHOT to 2.1-SNAPSHOT. I realise that the Java API hasn't actually changed, but since there is already a version 2.0.0.0 published on Maven Central, the next version published must be incrementally higher than that, at a minimum.

This revives hamcrest-integration as separate subproject. This was done by unzipping hamcrest-integration-1.3.jar downloaded from Maven Central. Unfortunately, the corresponding code available from the archived hamcrest-java on Google Code appeared out of date. I could not see test code for java-integration (I had a poke around the Google Code archive, but no luck there).

Salient points:
* New project hamcrest is the result of merging hamcrest-core and hamcrest-library.
* hamcrest-integration remains an independent project.
* hamcrest-integration remains only as a 1.3 line bugfix backstop. It is not published, and its version is pinned at 1.3.1-SNAPSHOT.
* All aggregated jars removed (no hamcrest-all or java-hamcrest), which means only a single artefact is published: org.hamcrest:hamcrest:<version> (jar, sources and javadoc)