Build empty jars for hamcrest-core and -library. (#236)
Existing maven builds that depend upon either `hamcrest-core-1.3.jar`
or `hamcrest-library.jar` will not upgrade their version for a pom-only
artifact. This change re-introduces a jar file for -core and -library,
but with no real classes. This allows Maven builds to correctly detect
the latest version of the jar. The transitive dependency on `hamcrest.jar`
brings in the latest version of the classes that were previously in
both `hamcrest-core.jar` or `hamcrest-library.jar`.