Add `MethodUtil` to make `hasProperty()` work for Java Records
The current `hasProperty()` matcher can't deal with
Java Records, as they are not compatible with JavaBeans
specification.

In this change, `HasProperty` first tries to do
the original thing and then tries to find a method
of which name is equivalent to the given property name,
if there's no matching property.

So for example,
if the class has a getter method named `property()`,
`hasProperty()` would regard it as the class has
a property named `property`.

I hope this change might be properly and easily removed
when the time comes and Hamcrest starts to support Java 17.