Implement matching conditions of `SamePropertyValuesAs` for Java Records
To achieve this goal I needed a different approach
from 8201577.
`samePropertyValuesAs(expectedBean, ignoredProperties)`
will recognize properties in java record classes if:

* There's a read accessor method of which name is same to the given property name

A read accessor method is a method
which acts like a getter for a field.
To check if the method is a read accessor method,
the algorithm filters the following:

1. The name of it should be identical to the name of the field.
2. It should return something.
3. It should not have any method parameters.

In this manner, we can now assure that
the collected methods are getters.

## Reference

* https://docs.oracle.com/en/java/javase/14/language/records.html